Canceled admin delete function

+++ app/Actions/Jetstream/DeleteUser.php
@@ -34,12 +34,14 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
-        DB::transaction(function () use ($user) {
-            $this->deleteTeams($user);
-            $user->deleteProfilePhoto();
-            $user->tokens->each->delete();
-            $user->delete();
-        });
+        if($user->email !== 'admin@tailadmin.dev'){
+            DB::transaction(function () use ($user) {
+                $this->deleteTeams($user);
+                $user->deleteProfilePhoto();
+                $user->tokens->each->delete();
+                $user->delete();
+            });
+        }
     }
 
     /**
